# api-spec.yaml
openapi: 3.0.3
info:
  title: Saitynai API
  version: 1.0.0
  description: |
    REST API for "Saitynai" community platform.
    Supports user registration, authentication, communities, posts, comments and admin operations (roles, blocking users).
servers:
  - url: https://saitynai-vl9u5.ondigitalocean.app/
    description: Production (DigitalOcean)
  - url: http://localhost:5000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /Community/{id}:
    get:
      tags: [Communities]
      summary: Get community by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Community data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityResponseDTO"
        "404":
          description: Community not found
    put:
      tags: [Communities]
      summary: Update community
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityDTO"
      responses:
        "200":
          description: Updated community
        "400":
          description: Invalid community data or ID
        "401":
          description: Not authenticated
        "403":
          description: Forbidden (not author or not Admin)
        "404":
          description: Community not found
        "422":
          description: Validation failed
    delete:
      tags: [Communities]
      summary: Delete community
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Community deleted
        "401":
          description: Not authenticated
        "403":
          description: Forbidden (not author or not Admin)
        "404":
          description: Community not found
...
  /Post/community/{comm_id}:
    get:
      tags: [Posts]
      summary: Get posts for a community
      security:
        - bearerAuth: []
      parameters:
        - name: comm_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of posts in community
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostResponseDTO"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
        "404":
          description: No posts found in community
  /Post/{id}:
    get:
      tags: [Posts]
      summary: Get post by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponseDTO"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
        "404":
          description: Post not found
    put:
      tags: [Posts]
      summary: Update post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdateDTO"
      responses:
        "200":
          description: Updated post
        "400":
          description: Invalid post data or ID
        "401":
          description: Not authenticated
        "403":
          description: Forbidden (not author or not Admin)
        "404":
          description: Post not found
        "422":
          description: Validation failed
    delete:
      tags: [Posts]
      summary: Delete post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Post deleted
        "401":
          description: Not authenticated
        "403":
          description: Forbidden (not author or not Admin)
        "404":
          description: Post not found
  /Post:
    post:
      tags: [Posts]
      summary: Create a new post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostCreateDTO"
            example:
              title: "First post"
              text: "Hello world"
              community_id: 1
      responses:
        "201":
          description: Post created
        "400":
          description: Invalid post data
        "401":
          description: Not authenticated
        "404":
          description: Community or user not found
        "422":
          description: Validation failed

  # ---------------- COMMENTS ----------------
  /Comment/post/{post_id}:
    get:
      tags: [Comments]
      summary: Get comments for a post
      security:
        - bearerAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentResponseDTO"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
        "404":
          description: No comments found for post
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CommunityDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    CommunityResponseDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        userId:
          type: integer
    PostCreateDTO:
      type: object
      properties:
        title:
          type: string
        text:
          type: string
        community_id:
          type: integer
    PostUpdateDTO:
      type: object
      properties:
        title:
          type: string
        text:
          type: string
    PostResponseDTO:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        text:
          type: string
        communityId:
          type: integer
        userId:
          type: integer
...
    CommentResponseDTO:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        postId:
          type: integer
        userId:
          type: integer